--- Packages/Netherlands3D/Core/Runtime/Scripts/Rendering.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0c017bd441cb90c438b4350ca65d4bc4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
