+++ Packages/Netherlands3D/Rendering/Runtime/Scripts/RendererHelpers.cs.meta
